+++ printf_lover_v2/utils/generated_tests/generated_tests_usr_version/includes/ft_printf.h
@@ -11,22 +11,29 @@
 #include <stdarg.h>
 #include <stdio.h>
 
-int g_count;
+int				g_count;
+typedef struct	_flag {
+	int		minus;
+	int		plus;
+	int		space;
+	int		zero;
+	int		dot;
+	int		width;
+	int		precision;
+	int		value;
+	int		v_len;
+	int		p_len;
+	int		sign;
+	char	*pvalue;
+}				Flag;
 
-typedef struct _flag {
-    int minus;
-    int plus;
-    int space;
-    int zero;
-    int dot;
-    int width;
-    int precision;
-    int value;
-    int v_len;
-    int p_len;
-    int sign;
-    char *pvalue;
-} Flag;
+void	print_per(va_list *ap, char *res);
+void	print_c(va_list *ap, char *res);
+void	print_s(va_list *ap, char *res);
+void	print_d(va_list *ap, char *res);
+void	print_p(va_list *ap, char *res);
+void	print_x(va_list *ap, char *res);
+void	print_u(va_list *ap, char *res);
 
 void	*ft_memset(void *b, int c, size_t len);
 void	*ft_calloc(size_t count, size_t size);
@@ -37,5 +44,16 @@ void	*ft_memcpy(void *d, void const *s, size_t n);
 size_t	ft_strlen(const char *src);
 char	*ft_itoa_base(long long value, int base, int is_upper);
 
-int ft_printf(const char *format, ...);
+void	set_flag(Flag *flag, va_list *ap, char *res, char type);
+
+void	put_str(char *a);
+void	print_value(Flag *flag);
+void	print_width(Flag *flag);
+void    print_in_order(Flag *flag, void (*a)(Flag *),void (*b)(Flag *), void (*c)(Flag *));
+
+void	set_res(va_list *ap, char *res);
+char	*set_dtarg(char *a, int *k);
+int		checker(va_list *ap, const char *a);
+void	print_rst(va_list *ap, const char *format);
+int		ft_printf(const char *format, ...);
 #endif
